chore: separate apply command to e2e folder
