+++ .circleci/config.yml
@@ -40,6 +40,7 @@ jobs:
           node-version: '18.15'
       - node/install-packages
       - run: npm run test-integration-with-reports
+      - run: npm run test-e2e-with-reports
       - store_test_results:
           path: reports
       - store_artifacts:
