+++ package.json
@@ -97,7 +97,9 @@
     "prepare": "husky install",
     "test": "npm run test-unit && npm run test-integration",
     "test-integration": "mocha --exit test/integration/**/*.test.ts",
+    "test-e2e": "mocha --exit test/e2e/**/*.test.ts",
     "test-integration-with-reports": "npm run test-integration -- --reporter mocha-junit-reporter --reporter-options mochaFile=reports/integration-results.xml",
+    "test-e2e-with-reports": "npm run test-e2e -- --reporter mocha-junit-reporter --reporter-options mochaFile=reports/e2e-results.xml",
     "test-unit": "mocha test/unit/**/*.test.ts",
     "test-unit-watch": "npm run test-unit -- -w --watch-files '**/*.ts'",
     "test-unit-with-reports": "npm run test-unit -- --reporter mocha-junit-reporter --reporter-options mochaFile=reports/unit-results.xml",
