--- test/integration/commands/apply/changesets/1-add-entry.json
@@ -1,39 +1,0 @@
-{
-  "sys": {
-    "type": "Changeset",
-    "createdAt": "1692107092473"
-  },
-  "items": [
-    {
-      "changeType": "add",
-      "entity": {
-        "sys": {
-          "type": "Link",
-          "linkType": "Entry",
-          "id": "4mfJMguVrMiQmZyeUdBbNu"
-        }
-      },
-      "data": {
-        "metadata": {
-          "tags": []
-        },
-        "sys": {
-          "id": "4mfJMguVrMiQmZyeUdBbNu",
-          "type": "Entry",
-          "contentType": {
-            "sys": {
-              "type": "Link",
-              "linkType": "ContentType",
-              "id": "author"
-            }
-          }
-        },
-        "fields": {
-          "name": {
-            "en-US": "Author 100"
-          }
-        }
-      }
-    }
-  ]
-}
